Se corrige problema del filtro de Oficina en la gestión de Turnos Desde el informe de Ocupación de Agendas ahora se puede abrir la agenda de la Oficina (en una nueva pestaña) a fin de facilitar el control de la misma Se hace un downgrade de jQuery 3.5 a 3.4.1 para que el menú se colapse en dispositivos móbiles Se ajusta package.json para anclar el componente datepicker a jQuery 3.4.1 porque sino no carga el plugin Se habilita todas las funciones de administración al Rol de Auditor. El panel de acciones se muestran para el rol editor o superior (no incluye al Auditor) Se deniega a nivel inferior al editor acciones en los distintos controladores para que no se puedan invocar desde la URL Se ponen mensajes de tipo flash a las acciones que se realizan desde la gestión de turnos Se agrega link en portada principal 'Más información sobre el sistema de turnos web para atención de profesionales y público en general' con enlace a documento

+++ src/Controller/LocalidadController.php
@@ -34,6 +34,9 @@ class LocalidadController extends AbstractController
      */
     public function new(Request $request): Response
     {
+        // Deniega acceso si no tiene un rol de editor o superior
+        $this->denyAccessUnlessGranted('ROLE_EDITOR');
+
         $localidad = new Localidad();
         $form = $this->createForm(LocalidadType::class, $localidad);
         $form->handleRequest($request);
@@ -67,6 +70,9 @@ class LocalidadController extends AbstractController
      */
     public function edit(Request $request, Localidad $localidad): Response
     {
+        // Deniega acceso si no tiene un rol de editor o superior
+        $this->denyAccessUnlessGranted('ROLE_EDITOR');
+
         $form = $this->createForm(LocalidadType::class, $localidad);
         $form->handleRequest($request);
 
@@ -87,6 +93,9 @@ class LocalidadController extends AbstractController
      */
     public function delete(Request $request, Localidad $localidad): Response
     {
+        // Deniega acceso si no tiene un rol de editor o superior
+        $this->denyAccessUnlessGranted('ROLE_EDITOR');
+
         if ($this->isCsrfTokenValid('delete'.$localidad->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($localidad);
@@ -101,6 +110,9 @@ class LocalidadController extends AbstractController
      */
     public function borraDiaAgendaTurnosbyLocalidad(Request $request, TurnoRepository $turnoRepository, OficinaRepository $oficinaRepository, Localidad $localidad, LoggerInterface $logger): Response
     {
+        // Deniega acceso si no tiene un rol de editor o superior
+        $this->denyAccessUnlessGranted('ROLE_EDITOR');
+
         //Construyo el formulario al vuelo
         $data = array(
             'fecha' => (new \DateTime(date("Y-m-d"))), // Campos del formulario
